Fixed: OFBiz Setup failing to setup the product store (OFBIZ-13211)
In a clan installation, the Product Store screen in OFBiz Setup fails to update.

The method createProductStoreWithDefaultSetting is expecting the table
ShipmentMethodType to be empty to insert the seed data from
OfbizSetupShippingData.xml.

But because OFBIZ-12608 introduced new seed data that inserts one record in
ShipmentMethodType, the condition expecting the table to be empty is not met
anymore and then the shipping data is not initialized.

Thanks: Carlos Guzman

jleroux: while at it update ofbiz-setup.adoc. We no longer use Ant.