Fixed: ClassCastException on PO Receipt in OFBiz (OFBIZ-13225)
When you realize a purchase order receipt the service receiveInventoryProduct return a success list with a message formatted by GString that raise an exception : 'GStringImpl cannot be cast to String'.

The service finish as success but the screen behind failed due to this exception.

To solve it we force the message as String

Thanks to Yashwant Dhakad for raise this issue