Fixed: Error viewing entities with fromDate key in entity maintenance (OFBIZ-13222)
Viewing entities with a fromDate as a primary seems to be broken with link like :
    https://demo-next.ofbiz.apache.org/webtools/control/entity/find/PartyContactMech/DemoCustomer/9015/2001-05-13%2000:00:00.0

This issue came from exception when we initialize java class java.net.URI with it, due to the presence of bad character %20.

To solve it, we encode the entity primary key path to escape this failure.

Thanks to Mickael Brohl for raise this issue and Jacques Leroux for the review