Fixed: Support for Add default browser security headers in writeJSONtoResponse (OFBIZ-13284) (#907)
This is to ensure SameSite cookies are sent on response.

Currently, the writeJSONtoResponse method correctly sends the JSON
response, but it does not include the SameSite attribute in the default
security headers. As a result, the SameSite cookie is not returned in
the browser response.

Expected Behavior:

writeJSONtoResponse should set default security headers for the
response.
SameSite attribute should be correctly applied to all cookies sent in
the response.
The browser should receive and respect the SameSite cookie.

Thanks: Sandeep Rajput